Agregado de metodos de user -> service -> DAOs -> controller y guardado de fotos

+++ src/services/users.services.js
@@ -17,7 +17,10 @@ class UsersService {
   // Obtiene un user dependiendo del user ID
   getUserById = async (uid) => {
     try {
-      const data = await query("SELECT id, name, last_name, email, username, photos, role FROM users WHERE id = $1", [uid]);
+      const data = await query(
+        "SELECT id, name, last_name, email, username, photos, role FROM users WHERE id = $1",
+        [uid]
+      );
       const user = data.rows[0];
       return user;
     } catch (err) {
@@ -27,7 +30,10 @@ class UsersService {
   // Obtiene un user dependiendo del user Email
   getUserByEmail = async (email) => {
     try {
-      const data = await query("SELECT id, name, last_name, email, username, photos, role FROM users WHERE email = $1", [email]);
+      const data = await query(
+        "SELECT id, name, last_name, email, username, photos, role FROM users WHERE email = $1",
+        [email]
+      );
       const user = data.rows[0];
       return user;
     } catch (err) {
@@ -48,11 +54,41 @@ class UsersService {
       return { error: true, data: err };
     }
   };
-  // Crea un usuario
+  // Actualiza un solo campo del user
+  updateUserField = async (uid, fieldName, fieldValue) => {
+    try {
+      const updateQuery = `UPDATE users SET ${fieldName} = $1 WHERE id = $2 RETURNING *`;
+      const userUpdated = await query(updateQuery, [fieldValue, uid]);
+      const updatedUser = userUpdated.rows[0];
+
+      if (!updatedUser) {
+        return { error: true, message: "No se pudo actualizar el usuario" };
+      }
+
+      return updatedUser;
+    } catch (err) {
+      return { error: true, data: err };
+    }
+  };
+  // Actualiza un usuario por su ID
+  updateUserById = async (uid, updatedUserData) => {
+    const { name, last_name, email, username, role, photos } = updatedUserData;
+    try {
+      const userUpdated = await query(
+        "UPDATE users SET name = $2, last_name = $3, email = $4,  role = $5, username = $6, photos = $7 WHERE id = $1 RETURNING id, email, name, last_name, username, photos, role",
+        [uid, name, last_name, email, role, username, photos]
+      );
+      const updatedUser = userUpdated.rows[0];
+      return updatedUser;
+    } catch (err) {
+      return { error: true, data: err };
+    }
+  };
+  // Elimina un usuario
   deleteUser = async (uid) => {
     try {
       const taskDeleted = await query(
-        "DELETE FROM users WHERE id = $1 RETURNING *",
+        "DELETE FROM users WHERE id = $1 RETURNING id, name, last_name, email, username, photos, role ",
         [uid]
       );
       return taskDeleted.rows[0];
